Revert "integrated the whole repo into single platform"
