+++ .github/workflows/ci.yml
@@ -1,5 +1,6 @@
 name: AI PR Comment Bot with Anthropic
 
+
 on:
   pull_request:
     types: [opened, synchronize]
@@ -7,15 +8,18 @@ on:
 jobs:
   comment:
     runs-on: ubuntu-latest
-
     steps:
       - name: Checkout repository
         uses: actions/checkout@v3
 
+    
       - name: Set up Python
         uses: actions/setup-python@v4
         with:
           python-version: '3.10'
+
+     
+      
 
       - name: Install dependencies
         run: |
@@ -24,8 +28,14 @@ jobs:
       - name: Print current working directory
         run: pwd
 
+      - name: List all files in the workspace
+        run: ls -l $GITHUB_WORKSPACE
+
+      - name: List files in .github/scripts
+        run: ls -l $GITHUB_WORKSPACE/.github/
+
       - name: Make the review script executable
-        run: chmod +x scripts/main.py
+        run: chmod +x $GITHUB_WORKSPACE/review_script.py
 
       - name: Check Python version
         run: python --version
@@ -37,4 +47,5 @@ jobs:
           REPO_OWNER: ${{ github.repository_owner }}
           REPO_NAME: ${{ github.event.repository.name }}
           PR_NUMBER: ${{ github.event.pull_request.number }}
-        run: python scripts/main.py
+        run: python $GITHUB_WORKSPACE/review_script.py
+
