--- evaluation_criteria/review_prompt.txt
@@ -1,10 +1,0 @@
-Review the following code changes:
-
-{{diff}}
-
-Please provide feedback on:
-- Code quality
-- Readability
-- Potential bugs
-- Optimization opportunities
-- Security or edge cases if applicable
