--- scripts/config.py
@@ -1,8 +1,0 @@
-import os
-
-# Constants
-GITHUB_TOKEN = os.getenv('GITHUB_TOKEN')
-ANTHROPIC_API_KEY = os.getenv('ANTHROPIC_API_KEY')
-REPO_OWNER = os.getenv('REPO_OWNER')
-REPO_NAME = os.getenv('REPO_NAME')
-PR_NUMBER = int(os.getenv('PR_NUMBER'))
